.java 파일로되어있던것, package 처리
