--- week1/강성욱/programmers/나머지가 1이 되는 수 찾기.java
@@ -1,10 +1,0 @@
-class Solution {
-	public int solution(int n) {
-		for (int i=1; i<n; i++) {
-			if (n % i == 1) {
-				return i;
-			}
-		}
-		return 0;
-	}
-}
