--- week1/강성욱/programmers/두 정수 사이의 합.java
@@ -1,13 +1,0 @@
-class Solution {
-    public long solution(int a, int b) {
-        int start = a < b ? a : b;
-        int end = a > b ? a : b;
-
-        long sum = 0L;
-        for(int i = start; i <= end; i++) {
-            sum += (long)i;
-        }
-
-        return sum;
-    }
-}
