+++ week1/강성욱/programmers/문자열_내_P와_Y의_개수.java
@@ -1,0 +1,21 @@
+package week1.강성욱.programmers;
+
+public class 문자열_내_P와_Y의_개수 {
+    class Solution {
+        boolean solution(String s) {
+            s = s.toLowerCase(); // 모두 소문자로 통일
+            int p = 0;
+            int y = 0;
+
+            for (char c : s.toCharArray()) {
+                if (c == 'p') {
+                    p++;
+                } else if (c == 'y') {
+                    y++;
+                }
+            }
+
+            return p == y;
+        }
+    }
+}
