--- week1/강성욱/programmers/문자열을 정수로 바꾸기.java
@@ -1,5 +1,0 @@
-class Solution {
-	public int solution(String s) {
-		return Integer.parseInt(s);
-	}
-}
