+++ week1/강성욱/programmers/문자열을_정수로_바꾸기.java
@@ -1,0 +1,9 @@
+package week1.강성욱.programmers;
+
+public class 문자열을_정수로_바꾸기 {
+    class Solution {
+        public int solution(String s) {
+            return Integer.parseInt(s);
+        }
+    }
+}
