--- week1/강성욱/programmers/자연수 뒤집어 배열로 만들기.java
@@ -1,19 +1,0 @@
-import java.util.*;
-
-class Solution {
-	public int[] solution(long n) {
-		List<Integer> list = new ArrayList<>();
-
-		while (n > 0) {
-			list.add((int)(n % 10));
-			n /= 10;
-		}
-
-		int[] result = new int[list.size()];
-		for (int i = 0; i < list.size(); i++) {
-			result[i] = list.get(i);
-		}
-
-		return result;
-	}
-}
