--- week1/강성욱/programmers/짝수와 홀수.java
@@ -1,5 +1,0 @@
-class Solution {
-	public String solution(int num) {
-		return num%2 == 0 ? "Even" : "Odd";
-	}
-}
