+++ week1/강성욱/programmers/짝수와_홀수.java
@@ -1,0 +1,9 @@
+package week1.강성욱.programmers;
+
+public class 짝수와_홀수 {
+    class Solution {
+        public String solution(int num) {
+            return num%2 == 0 ? "Even" : "Odd";
+        }
+    }
+}
