--- week1/강성욱/programmers/평균 구하기.java
@@ -1,7 +1,0 @@
-import java.util.Arrays;
-
-class Solution {
-    public double solution(int[] arr) {
-        return Arrays.stream(arr).average().orElse(0);
-    }
-}
