--- week1/강성욱/programmers/하샤드 수.java
@@ -1,13 +1,0 @@
-class Solution {
-	public boolean solution(int x) {
-		int sum = 0;
-		int temp = x;
-
-		while(temp > 0) {
-			sum += temp % 10;
-			temp /= 10;
-		}
-
-		return x%sum == 0 ? true : false;
-	}
-}
